jmx-scraper: reconnect when connection is lost
